replace the github command with the generate command, refactor
