+++ mono_repo/lib/src/commands/travis/generate.dart
@@ -1,93 +1,67 @@
-// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// Copyright (c) 2020, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-
-// ignore_for_file: directives_ordering
 
 import 'dart:io';
 
 import 'package:path/path.dart' as p;
 
-import '../ci_shared.dart';
-import '../ci_test_script.dart';
-import '../root_config.dart';
-import '../user_exception.dart';
-import 'github/github_yaml.dart';
-import 'mono_repo_command.dart';
-import 'travis.dart' show travisShPath;
+import '../../ci_shared.dart';
+import '../../ci_test_script.dart';
+import '../../root_config.dart';
+import '../../user_exception.dart';
+import 'travis_yaml.dart';
 
-const githubActionYamlPath = '.github/workflows/dart.yml';
+const travisFileName = '.travis.yml';
+const travisShPath = 'tool/travis.sh';
 
-class GitHubActionsCommand extends MonoRepoCommand {
-  @override
-  String get name => 'github';
-
-  @override
-  String get description => 'Configure GitHub Actions for child packages.';
-
-  GitHubActionsCommand() : super() {
-    argParser.addFlag(
-      'validate',
-      negatable: false,
-      help: 'Validates that the existing travis config is up to date with '
-          'the current configuration. Does not write any files.',
-    );
-  }
-
-  @override
-  void run() => generateGitHubActions(
-        rootConfig(),
-        validateOnly: argResults['validate'] as bool,
-      );
-}
-
-void generateGitHubActions(
+void generateTravisConfig(
   RootConfig rootConfig, {
   bool validateOnly = false,
 }) {
   validateOnly ??= false;
-  final githubConfig = _GeneratedGitHubConfig.generate(
+  final travisConfig = _GeneratedTravisConfig.generate(
     rootConfig,
   );
   if (validateOnly) {
     _validateFile(
       rootConfig.rootDirectory,
-      githubConfig.workflowYaml,
-      githubActionYamlPath,
+      travisConfig.travisYml,
+      travisFileName,
     );
     _validateFile(
       rootConfig.rootDirectory,
-      githubConfig.travisSh,
+      travisConfig.travisSh,
       travisShPath,
     );
   } else {
     writeFile(
       rootConfig.rootDirectory,
-      githubActionYamlPath,
-      githubConfig.workflowYaml,
+      travisFileName,
+      travisConfig.travisYml,
       isScript: false,
     );
     writeFile(
       rootConfig.rootDirectory,
       travisShPath,
-      githubConfig.travisSh,
+      travisConfig.travisSh,
       isScript: true,
     );
   }
 }
 
 /// The generated yaml and shell script content for travis.
-class _GeneratedGitHubConfig {
-  final String workflowYaml;
+class _GeneratedTravisConfig {
+  final String travisYml;
   final String travisSh;
 
-  _GeneratedGitHubConfig._(this.workflowYaml, this.travisSh);
+  _GeneratedTravisConfig._(this.travisYml, this.travisSh);
 
-  factory _GeneratedGitHubConfig.generate(RootConfig rootConfig) {
+  factory _GeneratedTravisConfig.generate(RootConfig rootConfig) {
     logPackages(rootConfig);
     final commandsToKeys = extractCommands(rootConfig);
 
-    final yml = generateGitHubYml(rootConfig, commandsToKeys);
+    final yml = generateTravisYml(rootConfig, commandsToKeys);
 
     final sh = generateTestScript(
       commandsToKeys,
@@ -95,7 +69,7 @@ class _GeneratedGitHubConfig {
       rootConfig.monoConfig.pubAction,
     );
 
-    return _GeneratedGitHubConfig._(yml, sh);
+    return _GeneratedTravisConfig._(yml, sh);
   }
 }
 
@@ -105,7 +79,7 @@ class TravisConfigOutOfDateException extends UserException {
   TravisConfigOutOfDateException()
       : super(
           'Generated travis config is out of date',
-          details: 'Rerun `mono_repo travis` to update generated config',
+          details: 'Rerun `mono_repo generate` to update generated config',
         );
 }
