+++ mono_repo/lib/src/runner.dart
@@ -8,7 +8,7 @@ import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
 
 import 'commands/check.dart';
-import 'commands/git_hub_actions.dart';
+import 'commands/generate.dart';
 import 'commands/mono_repo_command.dart';
 import 'commands/presubmit.dart';
 import 'commands/pub.dart';
@@ -18,7 +18,7 @@ import 'version.dart';
 final commands = List<Command<void>>.unmodifiable(
   [
     CheckCommand(),
-    GitHubActionsCommand(),
+    GenerateCommand(),
     PresubmitCommand(),
     PubCommand(),
     TravisCommand(),
