+++ mono_repo/test/readme_test.dart
@@ -5,7 +5,7 @@
 import 'dart:io';
 
 import 'package:mono_repo/src/ci_test_script.dart';
-import 'package:mono_repo/src/commands/travis.dart';
+import 'package:mono_repo/src/commands/travis/generate.dart';
 import 'package:mono_repo/src/package_config.dart';
 import 'package:test/test.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;
